Update to cdk 5.0.20120323. Implemented more of the API.
git-svn-id: http://maxmods.googlecode.com/svn/trunk@1358 b77917fb-863c-0410-9a43-030a88dac9d3

+++ curses.mod/src/cdk/draw.c
@@ -2,8 +2,8 @@
 
 /*
  * $Author: tom $
- * $Date: 2004/11/07 23:36:55 $
- * $Revision: 1.58 $
+ * $Date: 2011/05/15 19:58:29 $
+ * $Revision: 1.59 $
  */
 
 /*
@@ -13,26 +13,28 @@
 void initCDKColor (void)
 {
 #ifdef HAVE_START_COLOR
-   int color[]	= {
-			COLOR_WHITE,	COLOR_RED,	COLOR_GREEN,
-			COLOR_YELLOW,	COLOR_BLUE,	COLOR_MAGENTA,
-			COLOR_CYAN,	COLOR_BLACK
-		};
+   int color[] =
+   {
+      COLOR_WHITE, COLOR_RED, COLOR_GREEN,
+      COLOR_YELLOW, COLOR_BLUE, COLOR_MAGENTA,
+      COLOR_CYAN, COLOR_BLACK
+   };
    int pair = 1;
    int fg, bg;
    int limit;
 
-   if (has_colors()) {
-      start_color();
+   if (has_colors ())
+   {
+      start_color ();
 
       limit = (COLORS < 8) ? COLORS : 8;
 
       /* Create the color pairs. */
-      for (fg=0; fg < limit; fg++)
+      for (fg = 0; fg < limit; fg++)
       {
-	 for (bg=0; bg < limit; bg++)
+	 for (bg = 0; bg < limit; bg++)
 	 {
-	    init_pair (pair++, color[fg], color[bg]);
+	    init_pair ((short)pair++, (short)color[fg], (short)color[bg]);
 	 }
       }
    }
@@ -44,19 +46,19 @@ void initCDKColor (void)
  */
 void boxWindow (WINDOW *window, chtype attr)
 {
-   /* Set some variables. */
-   int tlx	= 0;
-   int tly	= 0;
-   int brx	= getmaxx(window) - 1;
-   int bry	= getmaxy(window) - 1;
+   /* *INDENT-EQLS* */
+   int tlx      = 0;
+   int tly      = 0;
+   int brx      = getmaxx (window) - 1;
+   int bry      = getmaxy (window) - 1;
 
    /* Draw horizontal lines. */
-   mvwhline(window, tly, 0, ACS_HLINE | attr, getmaxx(window));
-   mvwhline(window, bry, 0, ACS_HLINE | attr, getmaxx(window));
+   mvwhline (window, tly, 0, ACS_HLINE | attr, getmaxx (window));
+   mvwhline (window, bry, 0, ACS_HLINE | attr, getmaxx (window));
 
    /* Draw vertical lines. */
-   mvwvline(window, 0, tlx, ACS_VLINE | attr, getmaxy(window));
-   mvwvline(window, 0, brx, ACS_VLINE | attr, getmaxy(window));
+   mvwvline (window, 0, tlx, ACS_VLINE | attr, getmaxy (window));
+   mvwvline (window, 0, brx, ACS_VLINE | attr, getmaxy (window));
 
    /* Draw in the corners. */
    mvwaddch (window, tly, tlx, ACS_ULCORNER | attr);
@@ -70,28 +72,35 @@ void boxWindow (WINDOW *window, chtype attr)
  * This draws a box with attributes and lets the user define
  * each element of the box.
  */
-void attrbox (WINDOW *win, chtype tlc, chtype trc, chtype blc, chtype brc, chtype horz, chtype vert, chtype attr)
+void attrbox (WINDOW *win,
+	      chtype tlc,
+	      chtype trc,
+	      chtype blc,
+	      chtype brc,
+	      chtype horz,
+	      chtype vert,
+	      chtype attr)
 {
-   /* Set the coordinates. */
-   int x1	= 0;
-   int y1	= 0;
-   int y2	= getmaxy(win) - 1;
-   int x2	= getmaxx(win) - 1;
-   int count	= 0;
+   /* *INDENT-EQLS* */
+   int x1       = 0;
+   int y1       = 0;
+   int y2       = getmaxy (win) - 1;
+   int x2       = getmaxx (win) - 1;
+   int count    = 0;
 
    /* Draw horizontal lines. */
    if (horz != 0)
    {
-      mvwhline(win, y1, 0, horz | attr, getmaxx(win));
-      mvwhline(win, y2, 0, horz | attr, getmaxx(win));
+      mvwhline (win, y1, 0, horz | attr, getmaxx (win));
+      mvwhline (win, y2, 0, horz | attr, getmaxx (win));
       count++;
    }
 
    /* Draw vertical lines. */
    if (vert != 0)
    {
-      mvwvline(win, 0, x1, vert | attr, getmaxy(win));
-      mvwvline(win, 0, x2, vert | attr, getmaxy(win));
+      mvwvline (win, 0, x1, vert | attr, getmaxy (win));
+      mvwvline (win, 0, x2, vert | attr, getmaxy (win));
       count++;
    }
 
@@ -141,13 +150,13 @@ void drawObjBox (WINDOW *win, CDKOBJS *object)
 /*
  * This draws a line on the given window. (odd angle lines not working yet)
  */
-void drawLine  (WINDOW *window, int startx, int starty, int endx, int endy, chtype line)
+void drawLine (WINDOW *window, int startx, int starty, int endx, int endy, chtype line)
 {
-   /* De=clare some local vars. */
-   int xdiff	= endx - startx;
-   int ydiff	= endy - starty;
-   int x	= 0;
-   int y	= 0;
+   /* *INDENT-EQLS* */
+   int xdiff    = endx - startx;
+   int ydiff    = endy - starty;
+   int x        = 0;
+   int y        = 0;
 
    /* Determine if we are drawing a horizontal or vertical line. */
    if (ydiff == 0)
@@ -163,12 +172,13 @@ void drawLine  (WINDOW *window, int startx, int starty, int endx, int endy, chty
    else
    {
       /* We need to determine the angle of the line. */
-      int height	= xdiff;
-      int width		= ydiff;
-      int xratio	= (height > width ? 1 : (width / height));
-      int yratio	= (width > height ? (width / height) : 1);
-      int xadj		= 0;
-      int yadj		= 0;
+      /* *INDENT-EQLS* */
+      int height        = xdiff;
+      int width         = ydiff;
+      int xratio        = (height > width ? 1 : (width / height));
+      int yratio        = (width > height ? (width / height) : 1);
+      int xadj          = 0;
+      int yadj          = 0;
 
       /* Set the vars. */
       x = startx;
@@ -181,16 +191,16 @@ void drawLine  (WINDOW *window, int startx, int starty, int endx, int endy, chty
 	 /* Make the x and y adjustments. */
 	 if (xadj != xratio)
 	 {
-	    x	= (xdiff < 0 ? x-1 : x+1);
+	    x = (xdiff < 0 ? x - 1 : x + 1);
 	    xadj++;
 	 }
 	 else
 	 {
-	    xadj	= 0;
+	    xadj = 0;
 	 }
 	 if (yadj != yratio)
 	 {
-	    y	= (ydiff < 0 ? y-1 : y+1);
+	    y = (ydiff < 0 ? y - 1 : y + 1);
 	    yadj++;
 	 }
 	 else
@@ -208,17 +218,17 @@ void drawShadow (WINDOW *shadowWin)
 {
    if (shadowWin != 0)
    {
-      int x_hi = getmaxx(shadowWin) - 1;
-      int y_hi = getmaxy(shadowWin) - 1;
+      int x_hi = getmaxx (shadowWin) - 1;
+      int y_hi = getmaxy (shadowWin) - 1;
 
       /* Draw the line on the bottom. */
-      mvwhline(shadowWin, y_hi, 1, ACS_HLINE | A_DIM, x_hi);
+      mvwhline (shadowWin, y_hi, 1, ACS_HLINE | A_DIM, x_hi);
 
       /* Draw the line on the right. */
-      mvwvline(shadowWin, 0, x_hi, ACS_VLINE | A_DIM, y_hi);
+      mvwvline (shadowWin, 0, x_hi, ACS_VLINE | A_DIM, y_hi);
 
-      mvwaddch (shadowWin, 0,	 x_hi, ACS_URCORNER | A_DIM);
-      mvwaddch (shadowWin, y_hi, 0,    ACS_LLCORNER | A_DIM);
+      mvwaddch (shadowWin, 0, x_hi, ACS_URCORNER | A_DIM);
+      mvwaddch (shadowWin, y_hi, 0, ACS_LLCORNER | A_DIM);
       mvwaddch (shadowWin, y_hi, x_hi, ACS_LRCORNER | A_DIM);
       wrefresh (shadowWin);
    }
@@ -231,12 +241,12 @@ void writeBlanks (WINDOW *window, int xpos, int ypos, int align, int start, int 
 {
    if (start < end)
    {
-      unsigned want = (end - start) + 1000;
+      unsigned want = (unsigned)(end - start) + 1000;
       char *blanks = (char *)malloc (want);
 
       if (blanks != 0)
       {
-	 cleanChar (blanks, want - 1, ' ');
+	 cleanChar (blanks, (int)(want - 1), ' ');
 	 writeChar (window, xpos, ypos, blanks, align, start, end);
 	 freeChar (blanks);
       }
@@ -246,7 +256,13 @@ void writeBlanks (WINDOW *window, int xpos, int ypos, int align, int start, int 
 /*
  * This writes out a char * string with no attributes.
  */
-void writeChar (WINDOW *window, int xpos, int ypos, char *string, int align, int start, int end)
+void writeChar (WINDOW *window,
+		int xpos,
+		int ypos,
+		char *string,
+		int align,
+		int start,
+		int end)
 {
    writeCharAttrib (window, xpos, ypos, string, A_NORMAL, align, start, end);
 }
@@ -254,7 +270,14 @@ void writeChar (WINDOW *window, int xpos, int ypos, char *string, int align, int
 /*
  * This writes out a char * string with attributes.
  */
-void writeCharAttrib (WINDOW *window, int xpos, int ypos, char *string, chtype attr, int align, int start, int end)
+void writeCharAttrib (WINDOW *window,
+		      int xpos,
+		      int ypos,
+		      char *string,
+		      chtype attr,
+		      int align,
+		      int start,
+		      int end)
 {
    int display = end - start;
    int x;
@@ -262,19 +285,19 @@ void writeCharAttrib (WINDOW *window, int xpos, int ypos, char *string, chtype a
    if (align == HORIZONTAL)
    {
       /* Draw the message on a horizontal axis. */
-      display = MINIMUM(display, getmaxx(window) - 1);
-      for (x=0; x < display ; x++)
+      display = MINIMUM (display, getmaxx (window) - 1);
+      for (x = 0; x < display; x++)
       {
-	 mvwaddch (window, ypos, xpos+x, CharOf(string[x+start]) | attr);
+	 mvwaddch (window, ypos, xpos + x, CharOf (string[x + start]) | attr);
       }
    }
    else
    {
       /* Draw the message on a vertical axis. */
-      display = MINIMUM(display, getmaxy(window) - 1);
-      for (x=0; x < display ; x++)
+      display = MINIMUM (display, getmaxy (window) - 1);
+      for (x = 0; x < display; x++)
       {
-	 mvwaddch (window, ypos+x, xpos, CharOf(string[x+start]) | attr);
+	 mvwaddch (window, ypos + x, xpos, CharOf (string[x + start]) | attr);
       }
    }
 }
@@ -282,7 +305,13 @@ void writeCharAttrib (WINDOW *window, int xpos, int ypos, char *string, chtype a
 /*
  * This writes out a chtype * string.
  */
-void writeChtype (WINDOW *window, int xpos, int ypos, chtype *string, int align, int start, int end)
+void writeChtype (WINDOW *window,
+		  int xpos,
+		  int ypos,
+		  chtype *string,
+		  int align,
+		  int start,
+		  int end)
 {
    writeChtypeAttrib (window, xpos, ypos, string, A_NORMAL, align, start, end);
 }
@@ -290,28 +319,36 @@ void writeChtype (WINDOW *window, int xpos, int ypos, chtype *string, int align,
 /*
  * This writes out a chtype * string * with the given attributes added.
  */
-void writeChtypeAttrib (WINDOW *window, int xpos, int ypos, chtype *string, chtype attr, int align, int start, int end)
+void writeChtypeAttrib (WINDOW *window,
+			int xpos,
+			int ypos,
+			chtype *string,
+			chtype attr,
+			int align,
+			int start,
+			int end)
 {
-   int diff		= end - start;
-   int display		= 0;
-   int x		= 0;
+   /* *INDENT-EQLS* */
+   int diff             = end - start;
+   int display          = 0;
+   int x                = 0;
 
    if (align == HORIZONTAL)
    {
       /* Draw the message on a horizontal axis. */
-      display = MINIMUM(diff, getmaxx(window) - xpos);
-      for (x=0; x < display; x++)
+      display = MINIMUM (diff, getmaxx (window) - xpos);
+      for (x = 0; x < display; x++)
       {
-	 mvwaddch (window, ypos, xpos+x, string[x+start] | attr);
+	 mvwaddch (window, ypos, xpos + x, string[x + start] | attr);
       }
    }
    else
    {
       /* Draw the message on a vertical axis. */
-      display = MINIMUM(diff, getmaxy(window) - ypos);
-      for (x=0; x < display; x++)
+      display = MINIMUM (diff, getmaxy (window) - ypos);
+      for (x = 0; x < display; x++)
       {
-	 mvwaddch (window, ypos+x, xpos, string[x+start] | attr);
+	 mvwaddch (window, ypos + x, xpos, string[x + start] | attr);
       }
    }
 }
